lumera-crystal/backend/pyproject.toml
[project]
name = "lumera-crystal-api"
version = "0.1.0"
description = "LUMERA CRYSTAL backend API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "sqlalchemy>=2.0.36",
  "alembic>=1.14.0",
  "psycopg[binary]>=3.2.0",
  "pydantic-settings>=2.6.0",
  "email-validator>=2.2.0",
  "python-slugify>=8.0.4",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "python-multipart>=0.0.9",
  "pillow>=11.0.0",
  "moviepy>=1.0.3"
]

[tool.uv]
dev-dependencies = [
  "ruff>=0.7.0",
  "pytest>=8.3.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
